Adjust tests to respect purecap requirements (function pointer LSB and ambiguous provenance)

+++ testsuite/libffi.closures/cls_pointer.c
@@ -9,7 +9,7 @@
 
 void* cls_pointer_fn(void* a1, void* a2)
 {
-	void*	result	= (void*)((intptr_t)a1 + (intptr_t)a2);
+	void*	result	= (void*)((intptr_t)a1 + (long long)a2);
 
 	printf("0x%08x 0x%08x: 0x%08x\n", 
 	       (unsigned int)(uintptr_t) a1,
